feat(equation): add supplement option
